+++ src/System.Globalization/tests/StringInfo/StringInfoTests.cs
