+++ src/System.Globalization/tests/TextInfo/TextInfoTests.cs
